add stop alias and rename cnd -> conditions
